requirements_citrus.txt
accelerate==1.8.1
dacite==1.9.2
datasets==3.2.0
deepspeed==0.16.4
einops==0.8.1
fastapi==0.115.12
gradio==5.30.0
gradio_client==1.10.1
huggingface-hub==0.31.4
jieba==0.42.1
matplotlib==3.10.3
modelscope==1.26.0
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
omegaconf==2.3.0
openai==1.82.0
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.2.3
peft==0.15.1
pillow==11.0.0
pycocotools==2.0.10
PyYAML==6.0.2
qwen-vl-utils==0.0.11
requests==2.32.3
safetensors==0.5.3
sam2==1.1.0
simplejson==3.20.1
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
timm==1.0.16
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.52.4
transformers-stream-generator==0.0.5
trl==0.19.1
triton==3.2.0
vllm==0.8.2
wandb==0.21.0
xformers==0.0.29.post2
streamlit==1.48.0